test/burst-test/requirements.txt
pyyaml
prometheus-client